#40 Move doguFetcher into cesregistry package
