+++ controllers/cesregistry/testdata/redmine-descriptor-cm.yaml
