+++ controllers/cesregistry/testdata/simple-cas-dogu.json
@@ -1,0 +1,4 @@
+{
+  "Name": "official/cas",
+  "Version": "4.3.2-1"
+}
