+++ controllers/cesregistry/testdata/simple-nginx-dogu.json
@@ -1,0 +1,4 @@
+{
+  "Name": "official/nginx",
+  "Version": "1.2.3-4"
+}
