+++ controllers/cesregistry/testdata/simple-postfix-dogu.json
@@ -1,0 +1,4 @@
+{
+  "Name": "official/postfix",
+  "Version": "3.4.5-6"
+}
